FE-1529: Start the optimizer service from the website's dev task
`turbo run dev --filter @apps/petrinaut-website -- --with-optimizer-service` builds and starts the local Petrinaut Optimizer around the website's Vite server, replacing the root `dev:petrinaut-optimization` script and the standalone launcher. The service lifecycle lives in @local/petrinaut-optimizer-client's dev-service module so other dev tasks can reuse it.
