+++ .gitignore
@@ -155,7 +155,6 @@ seed-users.json
 AGENTS.local.md
 CLAUDE.local.md
 .claude/settings.local.json
-.claude/launch.json
 
 # Terraform
 **/.terraform
